Validation stringency is now respected in parsing of SAM file header.  Note that because the header is parsed in SAMFileReader ctor, the only way to control the validation stringency used to parse the header is to call SAMFileReader.setDefaultValidationStringency.
If stringency is not strict, validation errors are stored in the SAMFileHeader, so non-command-line clients (e.g. IGV) can access the errors programmatically.  

The text of the header as it appeared in the input is a property on SAMFileHeader.  Note that this is not a toString() operation.  Changes to the header after reading from the input are not reflected in this property.

@CO comment lines are now handled, stored in SAMFileHeader, merged in MergeSamFiles, and written to output SAM file.