Have ParsingUtils support FTP out-of-the-box
Add FTPHelper as an implementation of URLHelper, and RemoteURLHelper which delegates to HTTPHelper and FTPHelper as appropriate. ParsingUtils should now, by default, support FTP. Add additional tests to make sure of this. 

Clients can still register their own helper, improved the javadoc on registerHelperClass to make it clear what was expected.