The motivation for these changes is to speed up reading of VCFs, which use a slower custom line-reading class.  In order to use a different reader, it was necessary to make some changes to the the FeatureCodec API.
FeatureCodec had a tight coupling with PositionalBufferedStream, which was good and bad: it allowed the codec to query the position of the underlying stream at any time (necessary for indexing), but it also introduced a level of statefulness that prevented some optimizations.  Such optimizations might involve reading ahead in the stream (e.g., async reading), or using a different type of InputStream entirely.

With that in mind, I made the following changes to tribble:
* Abstract out FeatureCodec's dependency on PositionalBufferedStream, and instead have it depend upon any generic source class (such as a LineIterator).
* Make FeatureCodec capable of generating such a source given a generic InputStream. For the purposes of indexing, make FeatureCodec capable of generating an indexible source (see LocationAware), as well.  This latter source is likely slower than the non-indexible one, which is why it is generated via a different mechanism.
* Implementations for all the above API changes, especially in the area of AsciiFeatureCodec and using faster line readers.
* Some cleanup and generics-typing, while I was at it.