Fixed LiftoverVcf to properly handle Genotypes when the Alleles have been reverse complemented.
Apparently no one ever used this tool to lift over a VCF that had genotypes in it.
This fix is something that I pretty much just copied from the GATK version of the liftover.
However, it's important to fix it here in Picard since we are planning on deprecating the
GATK version in favor of maintaining just the Picard one.