targeted metrics now filters out low quality bases and enforces a minimum mapping quality
adding a median target coverage metric and the percentage of bases at 1x or greater to the targeted metrics.

do not count overlapping bases in targeted metrics.  To do this, we need
to clone the SAMRecord, so adding an option to allow side effects when
computing the targeted metrics.

fixes a bug where we were not counting bases from supplemental records
in targeted metrics