Added some code for comparing sequence data to fingerprints and cross checking it.
This code has been under development for a while - and has even been used in certain
production settings - but it is not yet ready for our wonderful users to actually use.
The code is not (yet) well tested and it uses some proprietary file formats that need
to go away in favor of htsjdk-supported ones.  Until we make those changes, we recommend
against using this code unless you really know what you are doing and are not using
sensitive data.

In other words, these tools are UNSUPPORTED (for now).  Do not ask for help (yet).