Refactored CollectWgsMetrics and added CollectWgsMetricsWithNonZeroCoverage.
Adding a tool to compute wgs metrics for all sites as well as for only
non-zero coverage sites: CollectWgsMetricsWithNonZeroCoverage.  This
tool produces a plot that gets produced based on the fraction of the
genome covered at a given mean fold coverage.  This plots the normalized
(by the mean fold coverage) cumulative coverage for both all sites and
non-zero sites, with the latter useful for when we do not have
full-coverage over all sites.

Refactored CollectWgsMetrics to create a metrics collector.