Adding a new tool for collecting metrics on independent replicates and Unique Molecular Identifiers.
The tool collects information regarding the relative amount of independent replicates (over hetrozygous sites of the sample) using replicate sets of size 2 and 3 (separately)
it also examine UMI's, if present and provides information regarding their performance with respect to the present alleles. Preliminary indications
www.broadinstitute.org/~farjoun/Posters/estimating-independent-replicates.pdf ) show that in several experiments there seems to be a correlation between the alleles even when the
UMI is different. This is highly non-intuitive and indicates that either experiments involving UMI's are harder to execute correclty than one would expect, or that writing code
that examines said UMI's is harder than expected...Either-way, I would be very happy if folks would collect some metrics on their favorite experiment with UMI and help me make this
a better tool that will QC UMI experiments well.

This is an alpha-class CLP. As you will see, there are many metrics, and I'm still not sure which of them are usefull. please do not expect the columns to remain constant for a while.