This commit enables MarkDuplicate to process query-sorted input.
Additionally, if the input is query-sorted it will mark the supplementary and secondary reads in the same manner as their
primary counter-parts (with the same queryname). Furthermore, in this case a unmapped read will be duplicate-marked as its
mapped mate (assuming, it is mated ofcourse).

Some tests had to be changed slghtly to accomodate for the different behaviours. tests have been added to test for the expected
functionality arround supplementary and secondary reads (which was untested), in both query and coordinate-sorted inputs.

**NOTA BENE:**
This commit includes an API change regarding the CLP parameter ASSUME_SORTED. this 'boolean' has been replaced
with a SortOrder parameter name ASSUME_ORDER which defualts to null (making the default behaviour the same). it enables
the user to force either query-sort or coordinate-sort behaviour.

Fixed tests and metric generation.