This commit enables MarkDuplicate to process query-sorted input. (#491)
Additionally, if the input is query-sorted it will mark the supplementary and secondary reads in the same manner as their
primary counter-parts (with the same queryname). Furthermore, in this case a unmapped read will be duplicate-marked as its
mapped mate (assuming, it is mated of course).

Some tests had to be changed slghtly to accomodate for the different behaviours. tests have been added to test for the expected
functionality around supplementary and secondary reads (which was untested), in both query and coordinate-sorted inputs.

**NOTA BENE:**

- Deprecated ASSUME_SORTED (replaced with ASSUME_SORT_ORDER)
- Fixed tests and metric generation.