Fix an error with too many open files for Multi-Tile BCLs.
When PerTileCycleParser seeks to a tile, it would call close() on the
underlying cycleFilesParser, which is fine if the parser parses only a
single tile.  When seeking to a new tile with MultiTileBclParser, we
want to r-use the underliny cycleFileParser, so the close() call on the
underlying iterator was commented out.  This leads to an error with "too
many open files", even when setting umilit to 800K files.

The solution is to only call close on the underlying iterator if the
base class' (PerTileCycleParser) close() method was called, not when the
seekToTile() method is called.