Adding a new tool for collecting metrics on independent replicates and UMIs (#482)
* Adding a new tool for collecting metrics on independent replicates and Unique Molecular Identifiers.

The tool collects information regarding the relative amount of independent replicates (over heterozygous sites of the sample) using replicate sets of size 2 and 3 (separately)
It also examine UMI's, if present, and provides information regarding their performance with respect to the present alleles. 

This is an alpha-class CLP. As you will see, there are many metrics, and I'm still not sure which of them are useful. Please do not expect the columns to remain constant for a while. And please feel free to suggest (via issues or PRs) enhancements and/or improvements!