Fix null library name problem in LibraryIdGenerator constructor (#505)
picard.sam.markduplicates.util.LibraryIdGenerator constructor creates new DuplicationMetrics and puts it to metricsByLibrary Map by key null. But when MarkDuplicates works it uses getLibraryName method that returns not null but "Unknown Library" value. As a result, when picard.sam.markduplicates.util.AbstractMarkDuplicatesCommandLineProgram::finalizeAndWriteMetrics method is called, there are 2 metrics by library in libraryIdGenerator: first id is null and it's empty, second is "Unknown Library" with some data. So in that case metric by unknown library and histogram will not be shown in metrics file.

The proposed fix uses "Unknown Library" instead of null String for library name.