ReorderSam was having trouble when new dictionary is larger than old one.
Also fixed logical problem that Reference was used both to open input (e.g. for cram) and to use as dictionary for new file.

Added a required input SEQUENCE_DICTIONARY and made REFERENCE_SEQUENCE not required.
This changes the API!

Added tests (was untested).

Refactored the code that writes a dictionary (using it in the test)

Fixes #974.

- responding to review comments
- fixing tests