Fix an issue where the google NIO code would try to load even when the NIO library wasn't present. (#1201)
* Fix an issue where the google NIO code would try to load even when the NIO library wasn't present.

Due to the way `GCS.isAvailable()` worked, a false positive could be generated when the google storage APIs were present, but the contrib NIO APIs were not. It turned out that the presence of the class `com.google.cloud.TransportOptions` would cause `isAvailable()` to return true.

This change requires that `initialize()` is successful, which is a better test of the presence of the NIO code.

It is slightly different than before in that it now initializes GCS immediately on startup, instead of waiting until `PathHelper.initilizeAll()` is called.