Allow explicit index file paths for inputs to CrosscheckFingerprints (#1823)
Add new INPUT_INDEX_MAP flag to allow a table of paths -> index files for use cases when an index is desired but not adjacent to the VCF/SAM files to be processed. Also added REQUIRE_INDEX_FILES flag to force the tool to only run when index files are either adjacent to the input files, or explicitly provided via a map. Tests were added to ensure the tool functions as desired.