Optionally extract all fingerprint snps (#1989)
Currently ExtractFingerprints only writes out the representative SNP for each haplotype block. While this is useful for the intended use-case (of creating a fingerprint and then using it subsequently in crosscheck or some such) there are other use-cases where it can be useful to know the genotype/PL at all the fingerprinting SNPs in each block.

Added a new argument EXTRACT_NON_REPRESENTATIVES_TOO to output all SNPs in each block.